backend/requirements.txt
fastapi==0.115.12
pydantic==2.11.3
uvicorn[standard]==0.34.2
psutil==7.0.0
httpx==0.28.1
asgi-lifespan==2.1.0
pytest==8.3.5
pytest-asyncio==0.26.0
mlx-lm==0.26.3
llama-cpp-python==0.3.8
